+++ Bridge/Sources/App/Coordinator/Coordinator.swift
@@ -13,12 +13,14 @@ protocol CoordinatorDelegate: AnyObject {
     func showSignInViewController()
     func showChannelViewController(of channel: Channel, navigationController: UINavigationController)
     func showProfileViewController(of userID: String, navigationController: UINavigationController)
+    func showWebPage(with urlString: URLString, navigationController: UINavigationController)
 }
 
 extension CoordinatorDelegate {
     func showSignInViewController() { }
     func showChannelViewController(of channel: Channel, navigationController: UINavigationController) { }
     func showProfileViewController(of userID: String, navigationController: UINavigationController) { }
+    func showWebPage(with urlString: URLString, navigationController: UINavigationController) { }
 }
 
 protocol Coordinator: AnyObject, Alertable {
